+++ package.json
@@ -63,6 +63,8 @@
     "@trpc/server": "^10.18.0",
     "chart.js": "^4.2.1",
     "chroma-js": "^2.4.2",
+    "csv-parse": "^5.3.6",
+    "csv-stringify": "^6.3.0",
     "d3-selection": "^3.0.0",
     "d3-zoom": "^3.0.0",
     "dotenv-cli": "^7.1.0",
@@ -102,8 +104,6 @@
     "@types/uuid": "^9.0.1",
     "@typescript-eslint/eslint-plugin": "^5.57.0",
     "@typescript-eslint/parser": "^5.57.0",
-    "csv-parse": "^5.3.6",
-    "csv-stringify": "^6.3.0",
     "dotenv": "^16.0.3",
     "eslint": "^8.36.0",
     "eslint-config-airbnb": "^19.0.4",
