+++ scripts/importUtilisateursIAM.ts
@@ -1,56 +1,7 @@
-/* eslint-disable import/no-extraneous-dependencies */
-import { parse } from 'csv-parse/sync';
-import { stringify } from 'csv-stringify/sync';
-import dotenv from 'dotenv';
-import { faker } from '@faker-js/faker/locale/fr';
-import { PrismaClient } from '@prisma/client';
 import process from 'node:process';
-import fs from 'node:fs';
 import assert from 'node:assert/strict';
 import logger from '@/server/infrastructure/logger';
-import {
-  créerUtilisateurs,
-  InputUtilisateur,
-  ProfilIdByCode,
-} from '@/server/infrastructure/accès_données/identité/seed';
-import { DIR_PROJET, DITP_ADMIN, DITP_PILOTAGE } from '@/server/domain/identité/Profil';
-
-// TODO: Must have
-// [x] installer le client keycloak
-// [x] créer le client keycloak sur dev-keycloak
-// [x] faire un get sur l'api keycloak
-// [x] faire fonctionner en ts ? abandonner le ts ?
-// [x] type: module = c'est quoi les conséquences ? on garde ? on jette ?
-// [x] faire un post pour créer un utilisateurs
-// [x] pouvoir lire les données dans un csv (quelle lib ?)
-// [x] quel format csv ?
-// [x] faire le ou les posts pour créer les utilisateurs
-// [x] générer les mots de passe temporaires
-// [x] écrire les mots de passe temporaires dans un log ou csv ou stdout pour pouvoir les communiquer
-//     autre option au début du traitement :
-//         - si l'entête du csv ne contient pas le champs 'Mot de passe', ajouter cette colonne et ré-écrire le csv
-//           source
-//         - sinon ne rien faire
-//         - poursuivre le traitement
-// [x] ajouter l'action 'doit reset son mot de passe' à la création de l'utilisateur
-// [x] documenter
-// [x] créer les utilisateurs dans l'app avec leur liste de chantiers
-// [x] supprimer version .mjs
-// [x] faire marcher si pas de chantier id (profils DITP Pilotage)
-// [ ] ajouter nom et prénom dans notre table utilisateur (on ne veut pas avoir à repasser sur les utilisateurs créés)
-// [ ] réfléchir 2 s à la gestion d'erreurs, qu'est-ce qui fait qu'on a confiance pour que ça marche ?
-// TODO: Nice to have
-// [x] besoin d'une confirmation car une fois les mots de passe générés ou affichés, si on relance la machine on écrase
-//     les valeurs et on les perd ?
-// [ ] si un utilisateur existe déjà chez nous, mettre à jour
-// [ ] vérifier qu'on n'a pas besoin du groupe dans keycloak, c'est une redondance des notions de profil et fonction.
-// [ ] convertir les records en données alignées au plus tôt pour ne plus avoir besoin de FIELDS au plus tôt.
-// [ ] envoie d'email aux utilisateurs créés ?
-// [ ] passer l'email à vérifié / non vérifié ?
-// [ ] ajouter une colonne status (utilisateur créé, utilisateur non créé, etc.) dans le csv à l'issu du script ?
-// [ ] comment on structure ça dans du code serveur admin ?
-// [ ] changer de devDependencies en dependencies pour le client kc ?
-// [ ] étendre à d'autres profils ?
+import ImportCsvUtilisateurs from '@/server/infrastructure/import_csv/identité/ImportCsvUtilisateurs';
 
 /**
  * Exemple de CSV :
@@ -94,196 +45,11 @@ import { DIR_PROJET, DITP_ADMIN, DITP_PILOTAGE } from '@/server/domain/identité
  * - ex de setup client ~ https://registry.terraform.io/providers/mrparkers/keycloak/latest/docs#keycloak-setup
  */
 
-dotenv.config();
-
-const CSV_PARSE_OPTIONS = {
-  columns: true,
-  skipEmptyLines: true,
-};
-const CSV_WRITE_OPTIONS = {
-  header:true,
-};
-
-const FIELDS: Record<string, string> = {
-  nom: 'Nom',
-  prénom: 'Prénom',
-  email: 'E-mail',
-  profils: 'Profils',
-  nomChantier: 'Nom du chantier',
-  idChantier: 'ID du chantier',
-  motDePasse: 'Mot de passe',
-};
-const EXPECTED_RECORD_FIELDS = Object.values(FIELDS);
-
-const KEYCLOAK_REALM = 'DITP';
-
-const CODES_PROFILS: Record<string, string> = {
-  ['DITP - Admin']: DITP_ADMIN,
-  ['DITP - Pilotage']: DITP_PILOTAGE,
-  ['Directeur de projet']: DIR_PROJET,
-};
-
-export type CsvRecord = Record<string, string>;
-
-function contientMotsDePasse(records: CsvRecord[]) {
-  const firstRecord = records[0];
-  return Boolean(firstRecord[FIELDS.motDePasse]);
-}
-
-function générerEtÉcrireMotsDePasse(records: CsvRecord[], filename: string) {
-  for (const record of records) {
-    assert(!record[FIELDS.motDePasse]);
-    // ex. de mot de passe généré : JHqVoxOoHnGLbhR
-    record[FIELDS.motDePasse] = faker.internet.password(15, false, /\w/, '');
-  }
-  const contents = stringify(records, CSV_WRITE_OPTIONS);
-  fs.writeFileSync(filename, contents);
-}
-
-export class ImportRecord {
-  constructor(
-    public readonly nom: string,
-    public readonly prénom: string,
-    public readonly email: string,
-    public readonly motDePasse: string,
-    public readonly profilCode: string,
-    public readonly chantierIds: string[],
-  ) {}
-
-  static fromCsvRecord(csvRecord: CsvRecord): ImportRecord {
-    const profilCode = CODES_PROFILS[csvRecord[FIELDS.profils]];
-
-    const chantierIds = [];
-    const csvChantierId = csvRecord[FIELDS.idChantier];
-    if (csvChantierId != '') {
-      chantierIds.push(csvChantierId);
-    }
-
-    return new ImportRecord(
-      csvRecord[FIELDS.nom],
-      csvRecord[FIELDS.prénom],
-      csvRecord[FIELDS.email],
-      csvRecord[FIELDS.motDePasse],
-      profilCode,
-      chantierIds,
-    );
-  }
-}
-
-export function parseCsvRecords(csvRecords: CsvRecord[]): ImportRecord[] {
-  assert(csvRecords, 'Erreur de parsing CSV. Pas de lignes ?');
-
-  const result: ImportRecord[] = [];
-  let lineNb = 0;
-  for (const csvRecord of csvRecords) {
-    lineNb += 1;
-    for (const field of EXPECTED_RECORD_FIELDS) {
-      assert.notEqual(csvRecord[field], null, `Erreur de parsing CSV ligne: ${lineNb}. Mauvais header ?`);
-    }
-    const nomDeProfil = csvRecord[FIELDS.profils];
-    assert(CODES_PROFILS[nomDeProfil], `Nom de profil ${nomDeProfil} inconnu. Profils connus : ${Object.keys(CODES_PROFILS)}`);
-    const importRecord = ImportRecord.fromCsvRecord(csvRecord);
-    result.push(importRecord);
-  }
-  return result;
-}
-
-// Oups?! Voir : https://github.com/TypeStrong/ts-node/discussions/1290
-// eslint-disable-next-line @typescript-eslint/no-implied-eval
-const dynamicImport = new Function('specifier', 'return import(specifier)');
-
-async function loginKcAdminClient() {
-  const { default: KcAdminClient } = await dynamicImport('@keycloak/keycloak-admin-client');
-  const keycloakUrl = process.env.IMPORT_KEYCLOAK_URL;
-  assert(keycloakUrl, 'Variable IMPORT_KEYCLOAK_URL manquante');
-  const kcAdminClient = new KcAdminClient({
-    baseUrl: keycloakUrl,
-    realmName: KEYCLOAK_REALM,
-    requestArgOptions: {},
-  });
-
-  const clientId = process.env.IMPORT_CLIENT_ID;
-  const clientSecret = process.env.IMPORT_CLIENT_SECRET;
-  if (!clientId || !clientSecret) {
-    throw new Error('Variable IMPORT_CLIENT_ID ou IMPORT_CLIENT_SECRET manquante.');
-  }
-
-  await kcAdminClient.auth({
-    grantType: 'client_credentials',
-    clientId,
-    clientSecret,
-  });
-
-  return kcAdminClient;
-}
-
-async function importeUtilisateurIAM(kcAdminClient: any, record: ImportRecord) {
-  const email = record.email;
-  const passwordCred = { temporary: true, type: 'password', value: record.motDePasse };
-  try {
-    await kcAdminClient.users.create({
-      realm: KEYCLOAK_REALM,
-      username: email,
-      email,
-      firstName: record.prénom,
-      lastName: record.nom,
-      enabled: true,
-      emailVerified: true,
-      requiredActions: ['UPDATE_PASSWORD'],
-      credentials: [passwordCred],
-    });
-    logger.info(`Utilisateur ${email} créé.`);
-  } catch (error: any) {
-    if (error.message == 'Request failed with status code 409') {
-      logger.warn(`L'email ${email} existe déjà.`);
-    } else {
-      logger.error(error);
-      throw error;
-    }
-  }
-}
-
-async function importeUtilisateursIAM(records: ImportRecord[]) {
-  const kcAdminClient = await loginKcAdminClient();
-  for (const record of records) {
-    await importeUtilisateurIAM(kcAdminClient, record);
-  }
-}
-
-async function importeUtilisateursPilote(records: ImportRecord[]) {
-  const inputUtilisateurs: InputUtilisateur[] = [];
-  for (const record of records) {
-    inputUtilisateurs.push({
-      email: record.email,
-      profilCode: record.profilCode,
-      chantierIds: record.chantierIds,
-    });
-  }
-
-  const prisma = new PrismaClient();
-  const resultSet = await prisma.profil.findMany({ select: { id: true, code: true } });
-  const profilIdByCode: ProfilIdByCode = {};
-  for (const profilRow of resultSet) {
-    profilIdByCode[profilRow.code] = profilRow.id;
-  }
-  await créerUtilisateurs(prisma, inputUtilisateurs, profilIdByCode);
-}
-
 async function main() {
   const filename = process.argv[2];
   assert(filename, 'Nom de fichier CSV manquant');
-
-  const contents = fs.readFileSync(filename, 'utf8');
-  const csvRecords: CsvRecord[] = parse(contents, CSV_PARSE_OPTIONS);
-
-  if (!contientMotsDePasse(csvRecords)) {
-    générerEtÉcrireMotsDePasse(csvRecords, filename);
-  }
-
-  const importRecords = parseCsvRecords(csvRecords);
-
-  await importeUtilisateursIAM(importRecords);
-  await importeUtilisateursPilote(importRecords);
+  const importateur = new ImportCsvUtilisateurs();
+  await importateur.importeFichierUtilisateurs(filename);
 }
 
 const isMain = eval('require.main === module');
