+++ src/server/domain/identité/UtilisateurIAMRepository.ts
@@ -1,0 +1,5 @@
+import UtilisateurPourIAM from '@/server/domain/identité/UtilisateurPourIAM';
+
+export interface UtilisateurIAMRepository {
+  ajouteUtilisateurs(utilisateurs: UtilisateurPourIAM[]): Promise<void>;
+}
