+++ src/server/domain/identité/UtilisateurPourIAM.ts
@@ -1,0 +1,8 @@
+export default class UtilisateurPourIAM {
+  constructor(
+    public readonly nom: string,
+    public readonly prénom: string,
+    public readonly email: string,
+    public readonly motDePasse: string,
+  ) {}
+}
